Update mealDoa to be more consistent with mealRepo

+++ app/src/main/java/com/github/se/polyfit/data/repository/MealRepository.kt
@@ -2,7 +2,9 @@ package com.github.se.polyfit.data.repository
 
 import android.content.Context
 import android.net.ConnectivityManager
+import com.github.se.polyfit.data.local.dao.MealDao
 import com.github.se.polyfit.data.remote.firebase.MealFirebaseRepository
+import com.github.se.polyfit.model.ingredient.Ingredient
 import com.github.se.polyfit.model.meal.Meal
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
@@ -13,29 +15,73 @@ import kotlinx.coroutines.withContext
 class MealRepository(
     private val context: Context,
     private val mealFirebaseRepository: MealFirebaseRepository,
-    private val dispatcher: CoroutineDispatcher = Dispatchers.Default
+    private val dispatcher: CoroutineDispatcher = Dispatchers.Default,
+    private val mealDao: MealDao
 ) {
-  private val checkConnecivity = connectivityChecker(context)
+    private val checkConnectivity = connectivityChecker(context)
+    private var isDataOutdated = false
 
-  suspend fun storeMeal(meal: Meal) =
-      withContext(dispatcher) { mealFirebaseRepository.storeMeal(meal).await() }
+    /**
+     * Stores a meal in the firebase and the local database. Makes sure both database are in sync
+     * @param meal the meal to store
+     */
+    suspend fun storeMeal(meal: Meal) {
+        if (checkConnectivity.checkConnection()) {
+            if (isDataOutdated) {
+                updateFirebase()
+                isDataOutdated = false
+            } else {
+                withContext(dispatcher) { mealFirebaseRepository.storeMeal(meal).await() }
+            }
+        } else {
+            isDataOutdated = true
+        }
+        mealDao.insert(meal)
+    }
 
-  suspend fun getMeal(firebaseID: String) =
-      withContext(dispatcher) { mealFirebaseRepository.getMeal(firebaseID).await() }
+    suspend fun getMeal(firebaseID: String): Meal? {
+        return mealDao.getMealByFirebaseID(firebaseID)
+    }
 
-  suspend fun getAllMeals() =
-      withContext(dispatcher) { mealFirebaseRepository.getAllMeals().await() }
+    suspend fun getAllMeals(): List<Meal?> {
+        return mealDao.getAllMeals()
+    }
 
-  suspend fun deleteMeal(firebaseID: String) =
-      withContext(dispatcher) { mealFirebaseRepository.deleteMeal(firebaseID).await() }
+    suspend fun deleteMeal(firebaseID: String) {
+        if (checkConnectivity.checkConnection()) {
+            if (isDataOutdated) {
+                updateFirebase()
+                isDataOutdated = false
+            }
+            withContext(dispatcher) { mealFirebaseRepository.deleteMeal(firebaseID).await() }
+
+        } else {
+            isDataOutdated = true
+        }
+        mealDao.deleteByFirebaseID(firebaseID)
+    }
+
+    suspend fun getAllIngredients(): List<Ingredient?> {
+        return mealDao.getAllIngredients()
+    }
+
+    private suspend fun updateFirebase() {
+        withContext(dispatcher) {
+            mealDao.getAllMeals().forEach {
+                if (it != null) {
+                    mealFirebaseRepository.storeMeal(it)
+                }
+            }
+        }
+    }
 }
 
 class connectivityChecker(private val context: Context) {
-  fun checkConnection(): Boolean {
-    val connectivityManager =
-        context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-    val currentNetwork = connectivityManager.activeNetwork
+    fun checkConnection(): Boolean {
+        val connectivityManager =
+            context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+        val currentNetwork = connectivityManager.activeNetwork
 
-    return currentNetwork == null
-  }
+        return currentNetwork == null
+    }
 }
